+++ apps/mobile/src/hooks/account.ts
@@ -11,27 +11,22 @@ import { Account, IPinAddress } from '@/core/services/preference';
 import { getWalletIcon } from '@/utils/walletInfo';
 import { filterMyAccounts } from '@/utils/account';
 import { useCreationWithShallowCompare } from './common/useMemozied';
-import {
-  accountEvents,
-  KeyringAccountWithAlias,
-} from '@/core/apis/account';
-import {
-  resolveValFromUpdater,
-  UpdaterOrPartials,
-} from '@/core/utils/store';
+import { accountEvents, KeyringAccountWithAlias } from '@/core/apis/account';
+import { resolveValFromUpdater, UpdaterOrPartials } from '@/core/utils/store';
 import balanceStore from '@/store/balance';
 import accountStore, {
   NEWLY_ADDED_ACCOUNT_DURATION,
   useAccountStore,
 } from '@/store/account';
-import { KeyringAccountWithAlias } from '@/core/apis/account';
-import { UpdaterOrPartials } from '@/core/utils/store';
 import { isSameAddress } from '@rabby-wallet/base-utils/dist/isomorphic/address';
-import { preferenceService } from '@/core/services';
+import { keyringService, preferenceService } from '@/core/services';
 import { EntityAccountBase } from '@/databases/entities/base';
 import { ormEvents } from '@/databases/entities/_helpers';
 import { InteractionManager } from 'react-native';
 import { appServiceEvents } from '@/core/services/_utils';
+import { Store } from '@/core/services/hdKeyringService';
+import { perfEvents } from '@/core/utils/perf';
+import { AccountInfoEntity } from '@/databases/entities/accountInfo';
 
 export type { KeyringAccountWithAlias as /** @deprecated */ KeyringAccountWithAlias };
 
@@ -141,99 +136,6 @@ export function useDevNewlyAddedAccounts() {
 
 export function startManageAccountStoreLifecycle() {
   accountStore.startLifecycle();
-  perfEvents.subscribe('USER_MANUALLY_UNLOCK', () => {
-    fetchAndSet();
-  });
-
-  keyringService.on('newAccount', (account: Account) => {
-    fetchAndSet();
-  });
-  // removedAccount
-  keyringService.on('removedAccount', async (account: Account) => {
-    fetchAndSet();
-
-    accountEvents.emit('ACCOUNT_REMOVED', {
-      removedAccounts: [account],
-    });
-
-    // Clean up backup reminder from preferenceService
-    const dbId = EntityAccountBase.buildDBId({
-      address: account.address,
-      type: account.type,
-      brandName: account.brandName,
-    });
-    preferenceService.clearNeedsBackupReminder(dbId);
-
-    await AccountInfoEntity.deleteByAccount(account);
-    await fetchNewlyAddedAccounts();
-  });
-
-  keyringService.store.subscribe(state => {
-    if (state.booted && state.vault) {
-      fetchAndSet();
-    }
-  });
-
-  accountEvents.on(
-    'ACCOUNT_ADDED',
-    async ({ accounts, scene, needsBackupReminder }) => {
-      // Store backup reminder in preferenceService (MMKV) for reliable persistence
-      if (needsBackupReminder) {
-        for (const account of accounts) {
-          const dbId = EntityAccountBase.buildDBId({
-            address: account.address,
-            type: account.type,
-            brandName: account.brandName,
-          });
-          preferenceService.setNeedsBackupReminder(dbId, true);
-        }
-      }
-      await AccountInfoEntity.recordNewAccount(accounts);
-      await fetchNewlyAddedAccounts();
-    },
-  );
-
-  ormEvents.on(`account_info:removed`, () => {
-    fetchNewlyAddedAccounts();
-  });
-
-  fetchNewlyAddedAccounts();
-  setInterval(() => {
-    InteractionManager.runAfterInteractions(() => {
-      fetchNewlyAddedAccounts();
-    });
-  }, 10 * 1e3);
-
-  setInterval(() => {
-    InteractionManager.runAfterInteractions(() => {
-      AccountInfoEntity.trimExpiredAccounts(NEWLY_ADDED_ACCOUNT_DURATION);
-    });
-  }, 60 * 1e3);
-}
-
-// 简单打个补丁先避免 balance 和 evmBalance 变化触发 ACCOUNTS_MAYBE_CHANGED
-// TODO: 彻底解决这个问题
-const stripAccountsForDiff = (accounts: KeyringAccountWithAlias[]) =>
-  accounts.map(account => {
-    const { balance, evmBalance, ...rest } = account;
-    return rest;
-  });
-function setAccounts(valOrFunc: UpdaterOrPartials<Store['accounts']>) {
-  zAccountStore.setState(prev => {
-    const { newVal, changed } = resolveValFromUpdater(
-      prev.accounts,
-      valOrFunc,
-      {
-        strict: true,
-      },
-    );
-
-    if (changed) {
-      return { ...prev, accounts: newVal };
-    }
-
-    return prev;
-  });
 }
 
 export function setCurrentAccount(
